move sparql-get from RDF layer to GET operation
