+++ src/lib/operationHandlers/readBlobHandler.ts
@@ -9,11 +9,58 @@ import { streamToObject, makeResourceData } from '../rdf/ResourceDataUtils'
 import { StoreManager } from '../rdf/StoreManager'
 import { resourceDataToRdf } from '../rdf/mergeRdfSources'
 import { rdfToResourceData } from '../rdf/rdfToResourceData'
-import { applyQuery } from '../rdf/applyQuery'
+import { newEngine } from '@comunica/actor-init-sparql-rdfjs'
+import { Store } from 'n3'
+
 import { getResourceDataAndCheckETag } from './getResourceDataAndCheckETag'
 import { ACL } from '../rdf/rdf-constants'
 
 const debug = Debug('read-blob-handler')
+
+function removeLeadingQuestionMark (str: string) {
+  return str.substring(1)
+}
+
+async function applyQuery (dataset: any, sparqlQuery: string): Promise<string> {
+  const store = new Store()
+  dataset.forEach((quad: any) => {
+    debug('quad', quad.toString())
+    store.addQuad(quad)
+  })
+  const myEngine = newEngine()
+  const result: any = await myEngine.query(sparqlQuery,
+    { sources: [ { type: 'rdfjsSource', value: store } ] })
+  // const sparqlResult = await myEngine.resultToString(result, 'application/json')
+  // debug(sparqlResult)
+  const bindings = await new Promise((resolve) => {
+    const bindings: any = []
+    result.bindingsStream.on('end', () => {
+      resolve(bindings)
+    })
+    result.bindingsStream.on('data', (data: any) => {
+      const binding: any = {}
+      const obj = JSON.parse(JSON.stringify(data))
+      debug(obj)
+      for (const key in obj) {
+        binding[removeLeadingQuestionMark(key)] = {
+          type: obj[key].termType.toLowerCase(),
+          value: obj[key].value
+        }
+      }
+      bindings.push(binding)
+    })
+  })
+  return JSON.stringify({
+    head: {
+      vars: result.variables.map(removeLeadingQuestionMark)
+    },
+    results: {
+      ordered: false,
+      distinct: false,
+      bindings
+    }
+  })
+}
 
 export const readBlobHandler = {
   canHandle: (wacLdpTask: WacLdpTask) => (wacLdpTask.wacLdpTaskType() === TaskType.blobRead),
